requirements.txt
progressbar2==3.47.0
python-gnupg==0.4.5
python-utils==2.3.0
six==1.12.0
